+++ tj-learning/src/main/java/com/tianji/learning/service/ILearningLessonService.java
@@ -32,5 +32,5 @@ public interface ILearningLessonService extends IService<LearningLesson> {
 
     Long queryCourseValid(Long courseId);
 
-    Long countLearningLessonByCourse(Long courseId);
+    Integer countLearningLessonByCourse(Long courseId);
 }
